Refactored map page
