+++ app/containers/MapPage/wraper.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  display: flex;
+  flex-direction: row;
+`;
